+++ python/tests/test_convert.py
@@ -84,6 +84,16 @@ nulls[0] = 11
 arr = d_arr.copy_to_host()
 print("itos(nulls=\\b1011):",nvstrings.itos(arr,nulls=nulls))
 
+# ltos
+arr = np.array([1.23, 123456789, -12.34, 987, 1234567890], dtype=np.int64)
+d_arr = rmm.to_device(arr)
+s = nvstrings.ltos(d_arr)
+print("ltos():",s)
+nulls = np.empty(int(s.size()/8)+1, dtype=np.int8)
+nulls[0] = 11
+arr = d_arr.copy_to_host()
+print("ltos(nulls=\\b1011):",nvstrings.ltos(arr,nulls=nulls))
+
 # ftos
 arr = np.array([1.23, 123456789, 0.00000987, -12.34, 9.87e+5, 6.54e-5, np.nan, np.inf], dtype=np.float32)
 d_arr = rmm.to_device(arr)
@@ -94,6 +104,17 @@ nulls[0] = 11
 s = nvstrings.ftos(arr,nulls=nulls)
 print("ftos(nulls=\\b1011):",s)
 
+# dtos
+arr = np.array([1.23, 1234567890, 0.00000987, -12.34, 9.87e+55, 6.54e-55, np.nan, np.inf], dtype=np.float64)
+d_arr = rmm.to_device(arr)
+s = nvstrings.dtos(d_arr)
+print("dtos:",s)
+nulls = np.empty(int(s.size()/8)+1, dtype=np.int8)
+nulls[0] = 11
+s = nvstrings.dtos(arr,nulls=nulls)
+print("dtos(nulls=\\b1011):",s)
+
+# IPv4
 s = nvstrings.to_device(["192.168.0.1","10.0.0.1",None,"","hello"])
 print(s)
 print(".ip2int()",s.ip2int())